requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cloudinary==1.22.0
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.1
django-contrib-comments==1.9.2
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-import-export==2.3.0
django-recaptcha==2.0.6
django-summernote==0.8.11.6
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.10
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0